--- src/worldofzuul/Plant.java
@@ -1,4 +1,0 @@
-package worldofzuul;
-
-public class Plant {
-}
