+++ src/worldofzuul/item/Harvester.java
@@ -1,0 +1,7 @@
+package worldofzuul.item;
+
+public class Harvester extends Item {
+    public Harvester(String name) {
+        super(name);
+    }
+}
