+++ src/worldofzuul/item/Pesticide.java
@@ -1,0 +1,4 @@
+package worldofzuul.item;
+
+public class Pesticide {
+}
