+++ src/worldofzuul/item/Seed.java
@@ -1,0 +1,7 @@
+package worldofzuul.item;
+
+public class Seed extends Item {
+    public Seed(String name) {
+        super(name);
+    }
+}
