+++ src/worldofzuul/parsing/CommandWords.java
@@ -1,4 +1,4 @@
-package worldofzuul;
+package worldofzuul.parsing;
 import java.util.HashMap;
