+++ src/worldofzuul/parsing/Parser.java
@@ -1,7 +1,6 @@
-package worldofzuul;
+package worldofzuul.parsing;
 
 import java.util.Scanner;
-import java.util.StringTokenizer;
 
 public class Parser 
 {
