+++ src/worldofzuul/world/Block.java
@@ -1,4 +1,4 @@
-package worldofzuul;
+package worldofzuul.world;
 
 public class Block extends GameObject {
 }
